+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/roadsigns/Dependencies.kt
@@ -29,7 +29,7 @@ object Libs {
     private object Versions {
         const val annotation = "1.5.0"
         const val appCompat = "1.4.1"
-        const val constraintLayout = "2.1.3"
+        const val constraintLayout = "2.1.4"
         const val espresso = "3.5.0"
         const val testExtJunit = "1.1.4"
         const val vectorDrawable = "1.1.0"