+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/roadsigns/Dependencies.kt
@@ -12,7 +12,7 @@ object Android {
 object Plugins {
 
     private object Versions {
-        const val android = "7.1.2"
+        const val android = "7.3.1"
         const val dokka = "1.7.20"
         const val kotlin = "1.7.21"
         const val versions = "0.44.0"