+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changes
 
+* Use Android plugin for Gradle v.7.3.1.
 * Use Kotlin v.1.7.21.
 * Use constraintlayout v.2.1.4.